+++ scripts/mdp_book_builder.py
@@ -11,9 +11,9 @@ import gzip
 import dpkt
 import binascii
 
-from orderbook import SecDef
-from orderbook import PacketProcessor
-from orderbook import ConsolePrinter
+from mdp.orderbook import SecDef
+from mdp.orderbook import PacketProcessor
+from mdp.orderbook import ConsolePrinter
 
 from sbedecoder import SBESchema
 from sbedecoder import SBEMessageFactory
