+++ setup.py
@@ -9,7 +9,7 @@ setup(
     license="MIT",
     keywords="sbe mdp3 orderbook",
     url="https://github.com/tfgm/sbedecoder",
-    packages=['sbedecoder', 'orderbook'],
+    packages=['sbedecoder', 'mdp'],
     scripts=['scripts/mdp_decoder.py', 'scripts/mdp_book_builder.py'],
     long_description='see https://github.com/tfgm/sbedecoder/INSTALL.md',
     install_requires=['dpkt', 'lxml', 'nose', 'mako', 'autopep8'],
